requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
fastapi==0.135.2
Flask==3.1.3
flask-cors==6.0.2
google_search_results==2.4.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kombu==5.6.2
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
msgpack==1.1.2
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai==2.28.0
packaging==26.0
pandas==3.0.0
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
prompt_toolkit==3.0.52
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311; sys_platform == "win32"
PyYAML==6.0.3
redis==7.4.0
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6


google-genai==1.16.0
